requirements.txt
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycodestyle==2.7.0
PyYAML==5.4.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
